+++ app/src/main/java/com/ezymd/restaurantapp/reviewsList/model/ShopReviewsBaseModel.kt
@@ -5,20 +5,20 @@ import java.util.*
 data class ShopReviewsBaseModel(var message: String?, var status: Int?, var data: RootData?)
 data class RootData(
     var listing: ArrayList<ShopRating>?,
-    var total: Int?,
-    var excellentRating: Int?,
-    var goodRating: Int?,
-    var averageRating: Int?,
-    var belowAverage: Int?,
+    var total: Double?,
+    var excellentRating: Double?,
+    var goodRating: Double?,
+    var averageRating: Double?,
+    var belowAverage: Double?,
     var totalReviews: Int?,
-    var poor: Int?
+    var poor: Double?
 )
 
 
 data class ShopRating(
     val order_id: Int,
-    val delivery_rating: Int,
-    val restaurant_rating: Int,
+    val delivery_rating: Double,
+    val restaurant_rating: Double,
     val feedback: String,
     val name: String,
     val image: String,
